Project still not working as of now. Still in transition in changing code for handling annotations. Only added some code to recognize keywords. Also started transition of integer types to more consistent integer types.

+++ general.c
@@ -149,7 +149,7 @@ void free_llist (void *list)
     }
 }
 
-void println (uint16_t no, char *buf)
+void println (unsigned no, char *buf)
 {
     printf("%6d: ", no);
     do {
